+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added endpoint `api_endpoint_exist()` to check if a API endpoint really exists
+
 ## [2.0.0] - 2025-09-30
 
 ### Added